instagram_mysite/requirements.txt
Django==5.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
daphne==4.0.0
uvicorn==0.24.0
